requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
Pillow>=10.0.0
PySide6>=6.5.0
scipy>=1.11.0
